feat: implement database trade history, dashboard view, startup scripts, and achieve 100% test coverage

+++ src/main/kotlin/com/gemini/krakenbot/service/KrakenService.kt
@@ -4,8 +4,8 @@ import com.gemini.krakenbot.model.OrderResult
 import com.gemini.krakenbot.model.TradeRecord
 import java.math.BigDecimal
 
-typealias RawBalances = Map<String, Double>
-typealias RawPrices = Map<String, Double>
+typealias RawBalances = Map<String, BigDecimal>
+typealias RawPrices = Map<String, BigDecimal>
 
 interface KrakenService {
     suspend fun getBalances(): RawBalances
